Fixed some compiler errors for MicroSoft Visual C++
git-svn-id: https://advants.svn.sourceforge.net/svnroot/advants@1046 fd2da0a4-d968-48df-9cb5-d8c496287cbf

+++ ImageRegistration/itkAvantsMutualInformationRegistrationFunction.cxx
@@ -527,7 +527,7 @@ AvantsMutualInformationRegistrationFunction<TFixedImage,TMovingImage,TDisplaceme
     {
       const double pRatio = vcl_log(jointPDFValue)-vcl_log(fixedImagePDFValue);
       term1 = dJPDF*pRatio;
-      term2 = vcl_log(2) * dFmPDF * jointPDFValue / fixedImagePDFValue ;
+      term2 = vcl_log((double)2) * dFmPDF * jointPDFValue / fixedImagePDFValue ;
       value =  (term2-term1);
     }  // end if-block to check non-zero bin contribution
   else value = 0;
@@ -566,7 +566,7 @@ AvantsMutualInformationRegistrationFunction<TFixedImage,TMovingImage,TDisplaceme
    {
     const double pRatio = vcl_log(jointPDFValue)-vcl_log(movingImagePDFValue);
     term1 = dJPDF*pRatio;
-    term2 = vcl_log(2) * dMmPDF * jointPDFValue / movingImagePDFValue ;
+    term2 = vcl_log((double)2) * dMmPDF * jointPDFValue / movingImagePDFValue ;
     value =  (term2-term1);
    }  // end if-block to check non-zero bin contribution
  else value = 0;
